requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzlocal==5.2
